+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/Communication.h
@@ -1,0 +1,17 @@
+#ifndef __COMMUNICATION_H__
+#define __COMMUNICATION_H__
+
+#include <Arduino.h>
+
+   void print_communication(); 
+   void initWiFi();
+   void connectWiFi();
+   void reset_wifi();
+
+   void initRadio(bool);
+   void checkDataOnRadio();
+   void publishData(String,bool);
+
+   char* string2char(String);
+  
+#endif
