+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/myCommon.cpp
@@ -1,0 +1,6 @@
+#include "Arduino.h"
+//#include "myCommon.h"
+
+void getMyMsg() {
+    Serial.println("from myCommon_lib");
+}
