+++ Code/ESP/WT_Energy_Monitoring_v2/lib/myCust_lib/myCommon.h
@@ -1,0 +1,8 @@
+#ifndef __MYCOMMON_H__
+#define __MYCOMMON_H__
+
+#include "Communication.h"
+
+  void getMyMsg();
+  
+#endif
